Refactored CDAMatchingEngine.
